Implementing Router
Implementing the ability to switch between tabs (ie. Search - Discovery): when pressing back, it will go back to the search screen before exit the application. If "is already on search screen" then it will just exit the application.
TODO: back button is not affecting the tab indicator of the bottom nav bar